refactor: consolidate MCPClient into MCPHttp
- Move error classification and retry logic to mcp_http.py
- Update MCPHttp.list_tools() to extract meta from tool responses
- Update mcps.py routes to use MCPConfig + MCPHttp (drop db dependency)
- Update deployment_service.py to use MCPHttp
- Delete overlapping mcp_client.py (848 lines)
